+++ sdk/keyvault/azure-security-keyvault-keys/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - Removed `Azure::Security::KeyVault::Keys::JsonWebKey::to_json`.
 
+### Bug Fixes
+
+- Fix getting a resume token from delete and recover key operations.
+
 ## 4.0.0-beta.1 (2021-04-07)
 
 ### New Features